# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto APIs
  version: 1.0.5
  description: >
    **"Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!"**
    
    WASAPhoto is web application based on RESTfull architecture to allows upload your photos directly from your PC, and they will be visible to everyone following you.
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  
  contact:
    name: Andrea Pizzi
    url: https://github.com/pizzi-andrea/WASAPhoto
    email: pizzi.1995517@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
    description: main server
  - url: http://localhost:4000
    description: debug server
paths:
...
  /users/{uid}/banned/{bannedId}:
    summary: users banned
    description: >
      this resource describe banned user
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/BannedId"
    
    put:
      tags:
        - user
        - ban
      summary: ban user identificated by *uid* 
      description: >
        taken uid of the user who wants to ban and uid of the user to be banned, 
        bans the user(BannedID)
              
      security:
        - BearerAuth: []
      
      operationId: banUser # required - used
      
      responses:
        "201":
          description: >
            success, banned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "204":
          description: >
            empty response, user just banned
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    delete:
      tags:
        - user
        - ban
      summary: unban user
      description: >
        taken uid of the user who wants unbband id of user banned and delete last one
      security:
        - BearerAuth: []
        
      operationId: unbanUser # required - used
      
      responses:
        "204":
          description: >
            success, unbanned user
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      tags:
        - user
        - ban
      summary: check if user had banned other user
      description: >
        taken uid of the user who wants check if banned. I if users had banned return it otherwise return empty response
      security:
        - BearerAuth: []
        
      operationId: isBanned # OK - used
      
      responses:
        "200":
          description: >
            success, user banned
          
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "204":
          description: >
            user not banned

        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/likes/{likeUserId}:
    
    summary: likes on photos
    description: >
      this endpoint rappresents likes putted on photo
    
    
    
    parameters:
        - $ref: "#/components/parameters/Uid"
        - $ref: "#/components/parameters/PhotoId"
        - $ref: "#/components/parameters/likeUserId"
    
    
    get:
      tags:
        - photo
        - like
      summary: check if user likeUserId has putted like on photo
      description: >
        Check if user identified by his uid has putted like on photo.
        If like has putted will return like otherwire return not found response
      
      operationId: checkLike # used
      
      security:
        - BearerAuth: []
      
      responses:
        "200":
          description: >
            success, like founded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: >
            not found, like not found
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
          
        
    
    put:
      
      tags:
        - photo
        - like
      summary: put like a photo
      description: >
        provide photo id and owner id. 
        The like is identified by the *uid* of the user who likes it.
      
      operationId: likePhoto # required - used
      
      security:
        - BearerAuth: []
      responses:
        "201":  
          description: >
            success, like added to the photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        
        "204": 
          description: >
            not content, just liked this photo
        
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          description: >
            Not found, user or photo not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
      
    delete:
      tags:
        - photo
        - like
      summary: remove like a photo
      description: >
        provide photo id and owner id
        and remove like given user identificated by
        *likeUserId*
        
      operationId: unlikePhoto # required - used
      security:
        - BearerAuth: []
      responses:
        "204":  
          description: >
            success, unliked the photo
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: >
            Not found, user or photo not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{uid}/myPhotos/{photoId}/comments/:
    description: >
      this endpoint rappresents a set of comments on photo
    summary: set comments on photo
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
    
    post:
      tags:
        - comment
        - photo
      summary: add comment a photo
      description: >
        give user id and photo id if user is owner of photo add a comment
        stored in the body of request
      
      
      
      requestBody:
        description: > 
          comment to add
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Comment"
      
      security:
        - BearerAuth: []
      operationId: commentPhoto # required - used
      
      responses:
        "201":
          description: >
            success, new comment added to photo
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Comment"
          
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: >
            Not found, photo or user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      security:
        - BearerAuth: []
        
      tags:
        - photo
        - comment
      summary: get comments on photo
      description: >
        give photo id and user id of owner photo then get all comments on photo
      
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
        - name: sortBy
          in: query
          required: false
          schema:
            description: >
              flag for sorting comments on photo by update data
            type: string
            enum: ["data"]
            default: "data"
          description: >
            sort comments by data creation
        
      
      operationId: getComments # used
      responses:
        "200":
          description: >
            success, get all comments on photo
          
          content:
            application/json:
              schema:
                description: >
                  list of comments on photo
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
                minItems: 1
                maxItems: 500

        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "204":
          description: >
            empty body, found nothing
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /session:
    summary: login action
    description: >
      endpoint that allow to perform login actions
    post:
      tags: 
        - login
      summary: Logs in the user
      description: >
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      
      operationId: doLogin # required - used
     
      requestBody:
        description: >
          User details
        content:
          application/json:
            schema:
              description: >
                username
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Username"
        required: true
      responses:
        "200":
          description: >
            User just registred, 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "201":
          description: >
            User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          $ref: "#/components/responses/BadRequest"
          
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    QueryLimit:
      description: >
        security parameter to limit the number of items in a response and avoid DOS attacks
      name: limit
      in: query
      required: false
      schema:
        description: >
          format limit values
        type: integer
        format: int
        minimum: 1
        example: 50
    UsernameFilter:
      description: >
        parameter use to filter set of users by his username
      name: username
      in: query
      required: false
      schema:
        description: >
          format username parameter
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 16
        example: "marioross"
    Uid:
      name: uid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        user id
    PhotoId:
      name: photoId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        photo id
...
    likeUserId:
      name: likeUserId
      in: path
      required: true
      description: >
        uid user put like
      schema:
        $ref: "#/components/schemas/Id"
...
    BannedId:
      name: bannedId
      in: path
      required: true
      description: >
        uid user to ban
      schema:
        $ref: "#/components/schemas/Id"
  responses:
    UnauthorizedError:
      description: >
        Access token is missing or invalid 
    UnauthorizedToken:
      description: >
        the submitted token does not have the necessary permissions to access this resource
    ServerError:
      description: >
        server internal error, for more information see log file
    BadRequest:
      description: >
        Bad request, invalid request format
    UserNotFound:
      description: >
        Not found, user not found
  securitySchemes:
    
    BearerAuth:
      description: >
        basic safety mechanism based on "bearer" mechanism. APIs that require a minimum level of security can adopt this method
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      description: >
        this object rappresent a user
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/Id"
        username: 
          $ref: "#/components/schemas/Username"
      required:
        - uid
        - username
    Username:
      description: >
        username of a user
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: "marioross"
    Id:
      description: >
        Identificator at 64-bit
      type: integer
      format: int64 
      readOnly: true
      example: 1
...
    TimeStamp:
      description: >
        this components describe timestamp value
        conform to [RFC3339](https://datatracker.ietf.org/doc/html/rfc3339#section-5.6) specification
      type: string
      format: data-time
      pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      example: "2017-07-21T17:32:28Z"
      minLength: 20
      maxLength: 21
      readOnly: true
    Token:
      description: > 
        Identification token. The token contains only one field with the user's uid
      type: object
      properties:
        value:
          $ref: "#/components/schemas/Id" # Id of owner
    Comment:
      description: > 
        this object rappresent a comment on a photo.
      type: object
      properties:
        commentId:
          $ref: "#/components/schemas/Id"
        author:
          $ref: "#/components/schemas/User"
  
        text:
          description: >
            comment text encoded in UNICODE format
          type: string
          minLength: 1
          maxLength: 250
          pattern: "^.*$"
          example: 
            "😀 i like you photo! 😀"
        timeStamp:
          $ref: "#/components/schemas/TimeStamp"
      required:
        - author
        - text
        - timeStamp
...